requirements.txt
streamlit
google-generativeai
yfinance
plotly
pandas
python-dotenv